REF: unify at2l0 + at1k4 implementations
