+++ solid_attenuator/sxr.py
@@ -10,10 +10,8 @@ This is intended to be used for the following attenuators:
 | AT2K2-SOLID | NEH 2.2    | H2.2 | 788.8 |
 | AT1K3-SOLID | TXI        | H1.1 | ~763  |
 """
-from typing import Dict, List
+from typing import Dict
 
-import numpy as np
-# from caproto import AlarmStatus
 from caproto.server import PVGroup, SubGroup
 from caproto.server.autosave import AutosaveHelper, RotatingFileManager
 from caproto.server.stats import StatusHelper
@@ -31,62 +29,6 @@ class SystemGroup(SystemGroupBase):
     This system group implementation is specific to AT2L0.
     """
 
-    async def motor_has_moved(self, blade_index, raw_state):
-        """
-        Callback indicating a motor has moved.
-
-        Update the current configuration, if necessary.
-
-        Parameters
-        ----------
-        blade_index : int
-            Blade index (not zero-based).
-
-        raw_state : int
-            Raw state value from control system.
-        """
-        array_idx = blade_index - self.parent.first_filter
-        state = State(int(raw_state))
-
-        new_config = list(self.active_config.value)
-        new_config[array_idx] = int(state)
-        if tuple(new_config) != tuple(self.active_config.value):
-            self.log.info('Active config changed: %s', new_config)
-            await self.active_config.write(new_config)
-            await self.active_config_bitmask.write(
-                util.int_array_to_bit_string(
-                    [State(blade).is_inserted for blade in new_config]
-                )
-            )
-            await self._update_active_transmission()
-
-        moving = list(self.filter_moving.value)
-        moving[array_idx] = state.is_moving
-        if tuple(moving) != tuple(self.filter_moving.value):
-            await self.filter_moving.write(moving)
-            await self.filter_moving_bitmask.write(
-                util.int_array_to_bit_string(moving)
-            )
-
-        flt = self.parent.filters[blade_index]
-        await flt.set_inserted_filter_state(state)
-
-    async def _update_active_transmission(self):
-        config = tuple(self.active_config.value)
-        offset = self.parent.first_filter
-        working_filters = self.parent.working_filters
-
-        transm = np.zeros_like(config) * np.nan
-        transm3 = np.zeros_like(config) * np.nan
-        for idx, filt in working_filters.items():
-            zero_index = idx - offset
-            if State(config[zero_index]).is_inserted:
-                transm[zero_index] = filt.transmission.value
-                transm3[zero_index] = filt.transmission_3omega.value
-
-        await self.transmission_actual.write(np.nanprod(transm))
-        await self.transmission_3omega_actual.write(np.nanprod(transm3))
-
     @util.block_on_reentry()
     async def run_calculation(self):
         energy = {
@@ -97,11 +39,7 @@ class SystemGroup(SystemGroupBase):
         desired_transmission = self.desired_transmission.value
         calc_mode = self.calc_mode.value
 
-        # This is a bit backwards - we reach up to the parent (IOCBase) for
-        # transmission calculations and such.
-        primary = self.parent
-
-        # material_check = primary.check_materials()
+        # material_check = self.check_materials()
         # await util.alarm_if(self.desired_transmission, not material_check,
         #                     AlarmStatus.CALC)
         # if not material_check:
@@ -114,14 +52,14 @@ class SystemGroup(SystemGroupBase):
 
         # Update all of the filters first, to determine their transmission
         # at this energy
-        for filter in primary.filters.values():
+        for filter in self.filters.values():
             await filter.set_photon_energy(energy)
 
         await self.calculated_transmission.write(
-            primary.calculate_transmission()
+            self.calculate_transmission()
         )
         await self.calculated_transmission_3omega.write(
-            primary.calculate_transmission_3omega()
+            self.calculate_transmission_3omega()
         )
 
         # Using the above-calculated transmissions, find the best configuration
@@ -129,14 +67,14 @@ class SystemGroup(SystemGroupBase):
         blade_transmissions = [
             [flt.transmission.value
              for flt in blade.active_filters.values()]
-            for blade in primary.filters.values()
+            for blade in self.filters.values()
         ]
 
         # Map per-blade array index -> filter index
         # Having removed non-active filters, these may not match 1-1 any more.
         blade_transmission_idx_to_filter_idx = [
             dict(enumerate(blade.active_filters))
-            for blade in primary.filters.values()
+            for blade in self.filters.values()
         ]
 
         config = calculator.get_ladder_config(
@@ -172,56 +110,6 @@ class SystemGroup(SystemGroupBase):
             config.filter_states,
         )
 
-    async def move_blade_step(self, state: Dict[int, State]):
-        """
-        Caller is requesting to move blades in or out.
-
-        The caller is expected to handle timeout scenarios and provide a
-        dictionary with which we can record this implementation's state.
-
-        Parameters
-        ----------
-        state : dict
-            State dictionary, which we use here to mark each time we request
-            a motion.  This will be passed in on subsequent calls.
-
-        Returns
-        -------
-        continue_ : bool
-            Returns `True` if there are more blades to move.
-        """
-        items = [
-            (pv, State(active), State(best)) for pv, active, best in
-            zip(
-                self._set_pvs, self.active_config.value, self.best_config.value
-            )
-        ]
-
-        move_out = {
-            pv: best
-            for pv, active, best in items
-            if not best.is_inserted and active != best
-        }
-        move_in = {
-            pv: best
-            for pv, active, best in items
-            if best.is_inserted and active != best
-        }
-
-        if move_in:
-            to_move = move_in
-            # Move blades IN first, to be safe
-        else:
-            to_move = move_out
-
-        for pv, target in to_move.items():
-            if state.get(pv, None) != target:
-                state[pv] = target
-                self.log.debug('Moving %s to %s', pv, target)
-                await self._pv_put_queue.async_put((pv, target))
-
-        return bool(move_in or move_out)
-
 
 class IOCBase(PVGroup):
     """
@@ -253,61 +141,6 @@ class IOCBase(PVGroup):
     autosave_helper = SubGroup(AutosaveHelper)
     stats_helper = SubGroup(StatusHelper, prefix=':STATS:')
     sys = SubGroup(SystemGroup, prefix=':SYS:')
-
-    @property
-    def working_filters(self):
-        """
-        A dictionary of all filters that are in working order.
-
-        That is to say, filters that are marked as active and not stuck.
-        """
-        return {
-            idx: filt for idx, filt in self.filters.items()
-            if filt.is_stuck.value != "True" and filt.active.value == "True"
-        }
-
-    def calculate_transmission(self):
-        """
-        Total transmission through all filter blades.
-
-        Stuck blades are assumed to be 'OUT' and thus the total transmission
-        will be overestimated (in the case any blades are actually stuck 'IN').
-        """
-        t = 1.
-        for filt in self.working_filters.values():
-            t *= filt.transmission.value
-        return t
-
-    def calculate_transmission_3omega(self):
-        """
-        Total 3rd harmonic transmission through all filter blades.
-
-        Stuck blades are assumed to be 'OUT' and thus the total transmission
-        will be overestimated (in the case any blades are actually stuck 'IN').
-        """
-        t = 1.
-        for filt in self.working_filters.values():
-            t *= filt.transmission_3omega.value
-        return t
-
-    @property
-    def all_transmissions(self):
-        """
-        List of the transmission values for working filters at the current
-        energy.
-
-        Stuck filters get a transmission of NaN, which omits them from
-        calculations/considerations.
-        """
-        T_arr = np.zeros(len(self.filters)) * np.nan
-        for idx, filt in self.working_filters.items():
-            T_arr[idx - self.first_filter] = filt.transmission.value
-        return T_arr
-
-    @property
-    def all_filter_materials(self) -> List[str]:
-        """All filter materials in a list."""
-        return [flt.material.value for flt in self.filters.values()]
 
 
 def create_ioc(prefix,
